Restore 32-bit build
This just made build possible. Functionality not tested.

Partially addresses #1195

+++ samples/sample_common/src/vaapi_utils.cpp
@@ -372,7 +372,7 @@ VAStatus CLibVA::AcquireVASurface(
     }
 
     AcquireCtx* ctx;
-    unsigned long handle=0;
+    uintptr_t handle=0;
     VAStatus va_res;
     VASurfaceAttrib attribs[2];
     VASurfaceAttribExternalBuffers extsrf;
